Unable to use float variables due to gcc bug (__clz_tab)
Working with a real distance sensors, ADC works, and also web pages

+++ avr-uip/apps/simple_ajax/simple_ajax.h
@@ -34,8 +34,7 @@ struct simple_httpd_state {
 	struct psock sockin, sockout;
 };
 
-
-extern uint16_t range;
+extern int range;
 
 
 void simple_httpd_appcall(void);
